See README for description

+++ src/main/c/qsolv.h
@@ -1,4 +1,4 @@
 #include "cunit.h"
 
 //prototype
-int solv(double a, double b, double c, double *x1, double *x2);
+double solv(double a, double b, double c, double *x1, double *x2);
